requirements.txt
pandas==2.2.2
openpyxl==3.1.2
python-pptx==0.6.23
reportlab==4.0.4
numpy==1.26.4
pdf2image==1.17.0
Pillow==10.3.0